chore: split instructor resolver
